Replace int by long int or std::size to manage larger amounts of values in matrices

+++ src/cpp/include/tbsla/cpp/MatrixCOO.hpp
@@ -13,7 +13,7 @@ class MatrixCOO : public virtual Matrix {
   public:
     friend std::ostream & operator<<( std::ostream &os, const MatrixCOO &m);
     MatrixCOO(int n_row, int n_col, std::vector<double> & values, std::vector<int> & row, std::vector<int> & col);
-    MatrixCOO(int n_row, int n_col, int n_values);
+    MatrixCOO(int n_row, int n_col, long int n_values);
     MatrixCOO(int n_row, int n_col);
     MatrixCOO() : values(0), row(0), col(0) {};
     std::vector<double> spmv(const std::vector<double> &v, int vect_incr = 0) const;
