+++ src/mpi/apps/tbsla_perf_page_rank_mpi.cpp
@@ -97,9 +97,9 @@ int main(int argc, char** argv) {
   MPI_Barrier(MPI_COMM_WORLD);
   auto t_op_end = now();
 
-  int sum_nnz = m->compute_sum_nnz(MPI_COMM_WORLD);
-  int min_nnz = m->compute_min_nnz(MPI_COMM_WORLD);
-  int max_nnz = m->compute_max_nnz(MPI_COMM_WORLD);
+  long int sum_nnz = m->compute_sum_nnz(MPI_COMM_WORLD);
+  long int min_nnz = m->compute_min_nnz(MPI_COMM_WORLD);
+  long int max_nnz = m->compute_max_nnz(MPI_COMM_WORLD);
 
   if(rank == 0) {
     auto t_app_end = now();
