+++ src/mpi/apps/tbsla_spmv_mpi.cpp
@@ -84,9 +84,9 @@ int main(int argc, char** argv) {
   std::vector<double> res = m->spmv(MPI_COMM_WORLD, vec);
   auto t_spmv_end = now();
 
-  int sum_nnz = m->compute_sum_nnz(MPI_COMM_WORLD);
-  int min_nnz = m->compute_min_nnz(MPI_COMM_WORLD);
-  int max_nnz = m->compute_max_nnz(MPI_COMM_WORLD);
+  long int sum_nnz = m->compute_sum_nnz(MPI_COMM_WORLD);
+  long int min_nnz = m->compute_min_nnz(MPI_COMM_WORLD);
+  long int max_nnz = m->compute_max_nnz(MPI_COMM_WORLD);
 
   if(rank == 0) {
     auto t_app_end = now();
